% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{flash_addin}
\alias{flash_addin}
\title{Run shinyFlash as an RStudio Addin}
\usage{
flash_addin(.data = NULL, path = NULL, width = 1000, height = 800)
}
\arguments{
\item{.data}{a data.frame containing columns `question` and `answer`}

\item{path}{the path to a `.xlsx`, `.csv`, `.rds` file containing a 
data.frame with columns `question` and `answer`}

\item{width}{the width of the rstudio modal dialog box (in pixels)}

\item{hieght}{the height of the rstudio modal dialog box (in pixels)}
}
\description{
Run shinyFlash as an RStudio Addin
}
