% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{flash_cards}
\alias{flash_cards}
\title{Run the Shiny Application}
\usage{
flash_cards(.data = NULL, path = NULL)
}
\arguments{
\item{.data}{a data.frame containing columns `question` and `answer`}

\item{path}{the path to a `.xlsx`, `.csv`, `.rds` file containing a 
data.frame with columns `question` and `answer`}

\item{...}{A series of options to be used inside the app.}
}
\description{
Run the Shiny Application
}
