% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{flash_data}
\alias{flash_data}
\alias{adv_r_deck}
\alias{drbc_deck}
\title{shinyFlash Datasets}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 32 rows and 3 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 326 rows and 3 columns.
}
\usage{
adv_r_deck

drbc_deck
}
\description{
Questions and Answers from the Advanced R book by Hadley Wickham 
(\link{https://adv-r.hadley.nz/}).

Questions and Answers about the Delaware River Basin Commission (DRBC). 
These questions and answers were amde from material taught by Carol Collier
and John Yagecic at Drexel University in the Winter Quarter of 2020.
}
\keyword{datasets}
