% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_flash_cards}
\alias{read_flash_cards}
\title{Read Flash card files}
\usage{
read_flash_cards(path)
}
\arguments{
\item{path}{the path to a file. Must have columns `question` and `answer`.}
}
\description{
Read Flash card files
}
\details{
function currently accepts .xlsx, .csv, and .rds
}
