% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{valid_flash_cards}
\alias{valid_flash_cards}
\title{validate that flash cards have the required columns}
\usage{
valid_flash_cards(.data)
}
\arguments{
\item{.data}{a date frame with columns `question` and `answer`}
}
\description{
validate that flash cards have the required columns
}
\details{
the column names are case insensitive
}
